move ConfirmationDIalog to commons/dialogs Folder
