+++ webclient/app/src/app/commons/dialog/ConfirmationDialog.jsx
